requirements.txt
Flask==3.1.2
h11==0.16.0
openai==2.8.1
python-dotenv==1.2.1